refactor(shared): rename BaseController file
