+++ src/shared/infra/http/helpers/ResponseHandler.ts
